feat(cvc-controller): remove cvc finalizer and update required logs
