+++ cmd/cstorvolume-claim/cvc/cstorvolumedeployment.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 The OpenEBS Authors
+Copyright 2019 The OpenEBS Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -238,6 +238,7 @@ func getContainerPort(port int32) []corev1.ContainerPort {
 	}
 }
 
+// createCStorTargetDeployment creates the cstor target deployment
 func createCStorTargetDeployment(
 	vol *apis.CStorVolume,
 ) (*appsv1.Deployment, error) {
